#pagamento-simplificado feat: alter migration type for amount

+++ database/migrations/2025_07_29_171942_create_wallets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->nullable()->index();
-            $table->integer('amount');
+            $table->float('amount', 8, 2);
             $table->timestamps();
         });
     }
